Extract db.setActiveSchemaV to define file
Cut-paste, no code has changed
