+++ internal/db/collection_define.go
@@ -17,6 +17,7 @@ import (
 
 	jsonpatch "github.com/evanphx/json-patch/v5"
 	"github.com/lens-vm/lens/host-go/config/model"
+	"github.com/sourcenetwork/immutable"
 
 	"github.com/sourcenetwork/defradb/client"
 	"github.com/sourcenetwork/defradb/client/request"
@@ -224,4 +225,149 @@ func (db *db) patchCollection(
 	}
 
 	return db.loadSchema(ctx)
+}
+
+// SetActiveSchemaVersion activates all collection versions with the given schema version, and deactivates all
+// those without it (if they share the same schema root).
+//
+// This will affect all operations interacting with the schema where a schema version is not explicitly
+// provided.  This includes GQL queries and Collection operations.
+//
+// It will return an error if the provided schema version ID does not exist.
+func (db *db) setActiveSchemaVersion(
+	ctx context.Context,
+	schemaVersionID string,
+) error {
+	if schemaVersionID == "" {
+		return ErrSchemaVersionIDEmpty
+	}
+	txn := mustGetContextTxn(ctx)
+	cols, err := description.GetCollectionsBySchemaVersionID(ctx, txn, schemaVersionID)
+	if err != nil {
+		return err
+	}
+
+	schema, err := description.GetSchemaVersion(ctx, txn, schemaVersionID)
+	if err != nil {
+		return err
+	}
+
+	colsWithRoot, err := description.GetCollectionsBySchemaRoot(ctx, txn, schema.Root)
+	if err != nil {
+		return err
+	}
+
+	colsBySourceID := map[uint32][]client.CollectionDescription{}
+	colsByID := make(map[uint32]client.CollectionDescription, len(colsWithRoot))
+	for _, col := range colsWithRoot {
+		colsByID[col.ID] = col
+
+		sources := col.CollectionSources()
+		if len(sources) > 0 {
+			// For now, we assume that each collection can only have a single source.  This will likely need
+			// to change later.
+			slice := colsBySourceID[sources[0].SourceCollectionID]
+			slice = append(slice, col)
+			colsBySourceID[sources[0].SourceCollectionID] = slice
+		}
+	}
+
+	for _, col := range cols {
+		if col.Name.HasValue() {
+			// The collection is already active, so we can skip it and continue
+			continue
+		}
+		sources := col.CollectionSources()
+
+		var activeCol client.CollectionDescription
+		var rootCol client.CollectionDescription
+		var isActiveFound bool
+		if len(sources) > 0 {
+			// For now, we assume that each collection can only have a single source.  This will likely need
+			// to change later.
+			activeCol, rootCol, isActiveFound = db.getActiveCollectionDown(ctx, colsByID, sources[0].SourceCollectionID)
+		}
+		if !isActiveFound {
+			// We need to look both down and up for the active version - the most recent is not necessarily the active one.
+			activeCol, isActiveFound = db.getActiveCollectionUp(ctx, colsBySourceID, rootCol.ID)
+		}
+
+		var newName string
+		if isActiveFound {
+			newName = activeCol.Name.Value()
+		} else {
+			// If there are no active versions in the collection set, take the name of the schema to be the name of the
+			// collection.
+			newName = schema.Name
+		}
+		col.Name = immutable.Some(newName)
+
+		_, err = description.SaveCollection(ctx, txn, col)
+		if err != nil {
+			return err
+		}
+
+		if isActiveFound {
+			// Deactivate the currently active collection by setting its name to none.
+			activeCol.Name = immutable.None[string]()
+			_, err = description.SaveCollection(ctx, txn, activeCol)
+			if err != nil {
+				return err
+			}
+		}
+	}
+
+	// Load the schema into the clients (e.g. GQL)
+	return db.loadSchema(ctx)
+}
+
+func (db *db) getActiveCollectionDown(
+	ctx context.Context,
+	colsByID map[uint32]client.CollectionDescription,
+	id uint32,
+) (client.CollectionDescription, client.CollectionDescription, bool) {
+	col, ok := colsByID[id]
+	if !ok {
+		return client.CollectionDescription{}, client.CollectionDescription{}, false
+	}
+
+	if col.Name.HasValue() {
+		return col, client.CollectionDescription{}, true
+	}
+
+	sources := col.CollectionSources()
+	if len(sources) == 0 {
+		// If a collection has zero sources it is likely the initial collection version, or
+		// this collection set is currently orphaned (can happen when setting migrations that
+		// do not yet link all the way back to a non-orphaned set)
+		return client.CollectionDescription{}, col, false
+	}
+
+	// For now, we assume that each collection can only have a single source.  This will likely need
+	// to change later.
+	return db.getActiveCollectionDown(ctx, colsByID, sources[0].SourceCollectionID)
+}
+
+func (db *db) getActiveCollectionUp(
+	ctx context.Context,
+	colsBySourceID map[uint32][]client.CollectionDescription,
+	id uint32,
+) (client.CollectionDescription, bool) {
+	cols, ok := colsBySourceID[id]
+	if !ok {
+		// We have reached the top of the set, and have not found an active collection
+		return client.CollectionDescription{}, false
+	}
+
+	for _, col := range cols {
+		if col.Name.HasValue() {
+			return col, true
+		}
+		activeCol, isFound := db.getActiveCollectionUp(ctx, colsBySourceID, col.ID)
+		if isFound {
+			return activeCol, isFound
+		}
+	}
+
+	return client.CollectionDescription{}, false
 }
